examples/nextjs-fullstack/backend/requirements.txt
auth-kit-fastapi==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # For PostgreSQL
# OR
# aiosqlite==0.19.0  # For SQLite